workers/requirements.txt
# requirements.txt
celery[gevent]
gevent
websocket-client
python-dotenv
requests

